requirements.txt
streamlit>=1.36
torch>=2.2
torchvision>=0.17
numpy>=1.24
Pillow>=10.0
